Rename matmul_splitK to matmul_splitK_webgpu
